pyproject.toml
[tool.poetry]
name = "python-steam-tradeoffer-manager"
version = "0.1.0"
license = "MIT"
description = ""
authors = ["SomeSpecialOne <Tkachenkodmitriy@yahoo.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
steamio = "^0.8.5"

[tool.poetry.dev-dependencies]
pytest = "^6.2.5"
pytest-asyncio = "0.16.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
